refactor(site): improve TaskStartupWarningButton API and file naming
- Renamed TaskStartupAlert.tsx to TaskStartupWarningButton.tsx to match exported component name
- Changed prop from 'task: Task' to 'lifecycleState: WorkspaceAgentLifecycle | null' for cleaner API
- Refactored conditional logic to use switch statement instead of if statements
- Updated TaskTopbar.tsx to pass lifecycleState directly
- Updated all Storybook stories to use new lifecycleState prop
- Removed unnecessary MockTask dependency from stories
